implemented getNonce function to generate and persist a random 10 digit number

+++ Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/user/models/User.kt
@@ -7,7 +7,7 @@ import javax.persistence.*
 @Table(schema = "public")
 data class User(@Id val publicWalletID: String,
                 val alias: String? = null,
-                var nonce: Int = 0,)
+                var nonce: Long = 0,)
 {
                 @ManyToMany(cascade = [CascadeType.ALL])
                 @JoinTable(
